+++ CHANGELOG.rst
@@ -12,7 +12,8 @@ new version (on deck)
 - [ENH] Add ``pivot_wider`` function, which is the inverse of the `pivot_longer`
   function. @samukweku
 - [INF] Add `openpyxl` to `environment-dev.yml`. @samukweku
-- [ENH] Add `as_categorical` method to create Categorical columns. @samukweku
+- [ENH] Add kwargs to ``encode_categorical`` function, to create ordered categorical columns,
+  or categorical columns with explicit categories. @samukweku
 
 v0.20.10
 =======